init/requirements.txt
azure-core==1.39.0
azure-storage-blob==12.28.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
cryptography==46.0.5
defusedxml==0.7.1
fonttools==4.62.1
fpdf2==2.8.7
idna==3.11
isodate==0.7.2
pillow==12.1.1
psycopg2-binary==2.9.11
pycparser==3.0
pydicom==3.0.2
requests==2.32.5
typing_extensions==4.15.0
urllib3==2.6.3
